Refactor test BUILD files into modular structure
- Add high-level fixture APIs for consistent testing patterns
- Rename test files to align with LSP protocol methods
- Consolidate redundant test files into feature-based organization
- Fix buildifier warnings in all BUILD.bazel files
