+++ test/slangd/core/BUILD.bazel
@@ -2,6 +2,8 @@
 Build file for slangd core tests.
 """
 
+load("@rules_cc//cc:cc_test.bzl", "cc_test")
+
 cc_test(
     name = "project_layout_service_test",
     timeout = "short",
