+++ test/slangd/semantic/definition_test.cpp
@@ -8,7 +8,6 @@
 #include <spdlog/spdlog.h>
 
 #include "../common/simple_fixture.hpp"
-#include "slangd/semantic/semantic_index.hpp"
 
 constexpr auto kLogLevel = spdlog::level::warn;
 
@@ -28,91 +27,7 @@ namespace slangd::semantic {
 
 using slangd::test::SimpleTestFixture;
 
-// Helper function to get consistent test URI
-inline auto GetTestUri() -> std::string {
-  return "file:///test.sv";
-}
-
-TEST_CASE("SemanticIndex provides O(1) symbol lookup", "[semantic_index]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    module test_module;
-      logic test_signal;
-      typedef logic [7:0] byte_t;
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-
-  fixture.AssertSymbolAtLocation(
-      *index, code, "test_signal", lsp::SymbolKind::kVariable);
-}
-
-TEST_CASE("SemanticIndex invalid location lookup", "[semantic_index]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    module simple;
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-
-  auto invalid_lookup = index->GetSymbolAt(slang::SourceLocation());
-  REQUIRE(!invalid_lookup.has_value());
-}
-
-TEST_CASE("SemanticIndex tracks references correctly", "[semantic_index]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    module test_module;
-      logic signal;
-      typedef logic [7:0] byte_t;
-
-      initial begin
-        signal = 1'b0;  // Reference to signal
-      end
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-
-  fixture.AssertReferenceExists(*index, code, "signal", 1);
-  fixture.AssertGoToDefinition(*index, code, "signal", 1, 0);
-}
-
-TEST_CASE("SemanticIndex basic symbol lookup", "[semantic_index]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    module test_module_unique;
-      logic test_signal;
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-
-  // Test basic symbol lookup functionality
-  fixture.AssertSymbolAtLocation(
-      *index, code, "test_signal", lsp::SymbolKind::kVariable);
-}
-
-TEST_CASE(
-    "SemanticIndex LookupDefinitionAt method exists and returns optional",
-    "[semantic_index]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    module test_module;
-      logic signal;
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-
-  auto result = index->LookupDefinitionAt(slang::SourceLocation());
-  REQUIRE(!result.has_value());
-}
-
-TEST_CASE(
-    "SemanticIndex module self-definition lookup works", "[semantic_index]") {
+TEST_CASE("SemanticIndex module self-definition lookup works", "[definition]") {
   SimpleTestFixture fixture;
   std::string code = R"(
     module empty_module;
@@ -121,12 +36,12 @@ TEST_CASE(
 
   auto index = fixture.CompileSource(code);
 
+  // Test go-to-definition: module name should resolve to itself
   fixture.AssertGoToDefinition(*index, code, "empty_module", 0, 0);
 }
 
 TEST_CASE(
-    "SemanticIndex parameter self-definition lookup works",
-    "[semantic_index]") {
+    "SemanticIndex parameter self-definition lookup works", "[definition]") {
   SimpleTestFixture fixture;
   std::string code = R"(
     module param_test;
@@ -137,12 +52,13 @@ TEST_CASE(
 
   auto index = fixture.CompileSource(code);
 
+  // Test go-to-definition: parameters should resolve to themselves
   fixture.AssertGoToDefinition(*index, code, "WIDTH", 0, 0);
   fixture.AssertGoToDefinition(*index, code, "ENABLE", 0, 0);
 }
 
 TEST_CASE(
-    "SemanticIndex typedef self-definition lookup works", "[semantic_index]") {
+    "SemanticIndex typedef self-definition lookup works", "[definition]") {
   SimpleTestFixture fixture;
   std::string code = R"(
     module typedef_test;
@@ -153,12 +69,12 @@ TEST_CASE(
 
   auto index = fixture.CompileSource(code);
 
+  // Test go-to-definition: typedefs should resolve to themselves
   fixture.AssertGoToDefinition(*index, code, "byte_t", 0, 0);
   fixture.AssertGoToDefinition(*index, code, "word_t", 0, 0);
 }
 
-TEST_CASE(
-    "SemanticIndex type cast reference lookup works", "[semantic_index]") {
+TEST_CASE("SemanticIndex type cast reference lookup works", "[definition]") {
   SimpleTestFixture fixture;
   std::string code = R"(
     module typecast_test;
@@ -173,6 +89,7 @@ TEST_CASE(
 
   auto index = fixture.CompileSource(code);
 
+  // Test go-to-definition: type cast reference should resolve to typedef
   fixture.AssertGoToDefinition(*index, code, "unique_cast_type", 1, 0);
 }
 
@@ -205,8 +122,7 @@ TEST_CASE(
 */
 
 TEST_CASE(
-    "SemanticIndex packed typedef parameter reference works",
-    "[semantic_index]") {
+    "SemanticIndex packed typedef parameter reference works", "[definition]") {
   SimpleTestFixture fixture;
   std::string code = R"(
     module test_packed;
@@ -216,12 +132,15 @@ TEST_CASE(
   )";
 
   auto index = fixture.CompileSource(code);
+
+  // Test go-to-definition: parameter reference in typedef should resolve to
+  // definition
   fixture.AssertGoToDefinition(*index, code, "PACKED_WIDTH", 1, 0);
 }
 
 TEST_CASE(
     "SemanticIndex unpacked typedef parameter go-to-definition",
-    "[semantic_index]") {
+    "[definition]") {
   SimpleTestFixture fixture;
   std::string code = R"(
     module test_unpacked_dims;
@@ -231,14 +150,16 @@ TEST_CASE(
   )";
 
   auto index = fixture.CompileSource(code);
+
+  // Test go-to-definition: parameter reference in unpacked dimensions should
+  // resolve to definition
   fixture.AssertGoToDefinition(*index, code, "ARRAY_SIZE", 1, 0);
 }
 
 TEST_CASE(
     "Parameter definition range should be name only, not full declaration",
-    "[semantic_index]") {
+    "[definition]") {
   SimpleTestFixture fixture;
-
   std::string code = R"(
     module test;
       parameter int WIDTH = 8;
@@ -261,6 +182,190 @@ TEST_CASE(
   auto range_length = result->end().offset() - result->start().offset();
 
   CHECK(range_length == 5);  // Now correctly returns just "WIDTH"
+}
+
+TEST_CASE(
+    "SemanticIndex reference tracking basic functionality", "[definition]") {
+  SimpleTestFixture fixture;
+  std::string code = R"(
+    module test_module;
+      logic signal;
+      typedef logic [7:0] byte_t;
+
+      initial begin
+        signal = 1'b0;  // Reference to signal
+      end
+    endmodule
+  )";
+
+  auto index = fixture.CompileSource(code);
+
+  // Test go-to-definition: signal reference should resolve to definition
+  fixture.AssertReferenceExists(*index, code, "signal", 1);
+  fixture.AssertGoToDefinition(*index, code, "signal", 1, 0);
+}
+
+TEST_CASE("SemanticIndex reference tracking in expressions", "[definition]") {
+  SimpleTestFixture fixture;
+  std::string code = R"(
+    module expression_test;
+      logic var_a, var_b, var_c;
+      logic [7:0] result;
+
+      always_comb begin
+        result = var_a ? var_b : var_c;
+        if (var_a && var_b) begin
+          result = 8'hFF;
+        end
+      end
+    endmodule
+  )";
+
+  auto index = fixture.CompileSource(code);
+
+  // Test that references are tracked in expressions
+  const auto& refs = index->GetReferences();
+  REQUIRE(!refs.empty());
+
+  // Test that we can find definitions for variables used in expressions
+  SimpleTestFixture::AssertContainsSymbols(
+      *index, {"var_a", "var_b", "var_c", "result"});
+}
+
+TEST_CASE(
+    "SemanticIndex handles interface references in expressions",
+    "[definition]") {
+  SimpleTestFixture fixture;
+  std::string code = R"(
+    module generic_module(generic_if iface);
+      logic state;
+      logic [7:0] counter;
+      logic enable;
+
+      always_comb begin
+        if (enable & ~iface.ready) begin
+          state = 1'b0;
+        end else if (enable & iface.ready) begin
+          if (iface.mode == 1'b1) begin
+            state = 1'b1;
+          end else begin
+            counter = iface.data;
+          end
+        end
+      end
+    endmodule
+  )";
+
+  auto index = fixture.CompileSource(code);
+
+  // Test that references through undefined interface members don't crash
+  // and that defined symbols are still captured
+  REQUIRE(index != nullptr);
+  REQUIRE(index->GetSymbolCount() > 0);
+  SimpleTestFixture::AssertContainsSymbols(
+      *index, {"state", "counter", "enable"});
+}
+
+TEST_CASE(
+    "SemanticIndex LookupDefinitionAt method basic functionality",
+    "[definition]") {
+  SimpleTestFixture fixture;
+  std::string code = R"(
+    module test_module;
+      logic signal;
+    endmodule
+  )";
+
+  auto index = fixture.CompileSource(code);
+
+  // Test that invalid location returns no result
+  auto result = index->LookupDefinitionAt(slang::SourceLocation());
+  REQUIRE(!result.has_value());
+}
+
+TEST_CASE(
+    "SemanticIndex collects definition ranges correctly", "[definition]") {
+  SimpleTestFixture fixture;
+  std::string code = R"(
+    module test_module;
+      logic signal;
+      typedef logic [7:0] byte_t;
+
+      initial begin : init_block
+        signal = 1'b0;
+      end
+    endmodule
+  )";
+
+  auto index = fixture.CompileSource(code);
+
+  // Verify symbols have definition ranges and is_definition flags set
+  const auto& all_symbols = index->GetAllSymbols();
+  REQUIRE(!all_symbols.empty());
+
+  // Verify expected symbols are present
+  SimpleTestFixture::AssertContainsSymbols(
+      *index, {"test_module", "signal", "byte_t", "init_block"});
+
+  // Test that all symbols have valid definition ranges
+  // This validates that the definition indexing is working correctly
+  bool found_at_least_one_definition = false;
+  for (const auto& [location, info] : all_symbols) {
+    if (info.is_definition) {
+      REQUIRE(info.definition_range.start().valid());
+      found_at_least_one_definition = true;
+    }
+  }
+  REQUIRE(found_at_least_one_definition);
+}
+
+TEST_CASE("SemanticIndex definition API compatibility", "[definition]") {
+  SimpleTestFixture fixture;
+  std::string code = R"(
+    module test_module;
+      logic signal;
+      typedef logic [7:0] byte_t;
+    endmodule
+  )";
+
+  auto index = fixture.CompileSource(code);
+
+  // Test reference storage API
+  const auto& references = index->GetReferences();
+  const auto& all_symbols = index->GetAllSymbols();
+
+  // Basic sanity checks - should have some data
+  REQUIRE(!all_symbols.empty());
+
+  // Verify references are accessible via GetReferences()
+  (void)references;  // May be empty for single-file tests
+
+  // Test that symbols have definition ranges in their SymbolInfo
+  // Verify at least one symbol has valid definition ranges
+  bool found_symbol_with_range = false;
+  for (const auto& [loc, info] : all_symbols) {
+    if (info.is_definition && info.location.valid()) {
+      found_symbol_with_range = true;
+      // Basic check that definition_range is set
+      REQUIRE(info.location.valid());
+      break;
+    }
+  }
+  REQUIRE(found_symbol_with_range);
+}
+
+TEST_CASE("SemanticIndex invalid location lookup", "[definition]") {
+  SimpleTestFixture fixture;
+  std::string code = R"(
+    module simple;
+    endmodule
+  )";
+
+  auto index = fixture.CompileSource(code);
+
+  // Test error handling: invalid location should return no result
+  auto invalid_lookup = index->GetSymbolAt(slang::SourceLocation());
+  REQUIRE(!invalid_lookup.has_value());
 }
 
 }  // namespace slangd::semantic
