--- test/slangd/semantic/type_handling_test.cpp
@@ -1,195 +1,0 @@
-#include <algorithm>
-#include <cstdlib>
-#include <memory>
-#include <string>
-#include <vector>
-
-#include <catch2/catch_all.hpp>
-#include <slang/ast/Symbol.h>
-#include <slang/util/Enum.h>
-#include <spdlog/spdlog.h>
-
-#include "../common/simple_fixture.hpp"
-#include "slangd/semantic/semantic_index.hpp"
-
-constexpr auto kLogLevel = spdlog::level::warn;
-
-auto main(int argc, char* argv[]) -> int {
-  spdlog::set_level(kLogLevel);
-  spdlog::set_pattern("[%l] %v");
-
-  // Suppress Bazel test sharding warnings
-  setenv("TEST_SHARD_INDEX", "0", 0);
-  setenv("TEST_TOTAL_SHARDS", "1", 0);
-  setenv("TEST_SHARD_STATUS_FILE", "", 0);
-
-  return Catch::Session().run(argc, argv);
-}
-
-namespace slangd::semantic {
-
-using slangd::test::SimpleTestFixture;
-
-// Helper function to get consistent test URI
-inline auto GetTestUri() -> std::string {
-  return "file:///test.sv";
-}
-
-TEST_CASE("SemanticIndex handles enum and struct types", "[type_handling]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    interface test_if;
-      logic clk;
-      logic rst;
-      modport master (input clk, output rst);
-    endinterface
-
-    module test_module(
-      test_if.master bus
-    );
-      typedef enum logic [1:0] {
-        IDLE,
-        ACTIVE,
-        DONE
-      } state_t;
-
-      state_t state;
-
-      typedef struct {
-        logic [7:0] data;
-        logic valid;
-      } packet_t;
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-
-  // Test LSP API: GetDocumentSymbols should return expected types
-  auto document_symbols = index->GetDocumentSymbols(GetTestUri());
-  REQUIRE(!document_symbols.empty());
-
-  // Check for interface with modport
-  SimpleTestFixture::AssertDocumentSymbolExists(
-      document_symbols, "test_if", lsp::SymbolKind::kInterface);
-  SimpleTestFixture::AssertDocumentSymbolExists(
-      document_symbols, "test_module", lsp::SymbolKind::kClass);
-}
-
-TEST_CASE(
-    "SemanticIndex collects definition ranges correctly", "[type_handling]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    module test_module;
-      logic signal;
-      typedef logic [7:0] byte_t;
-
-      initial begin : init_block
-        signal = 1'b0;
-      end
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-
-  // Verify symbols have definition ranges and is_definition flags set
-  const auto& all_symbols = index->GetAllSymbols();
-  REQUIRE(!all_symbols.empty());
-
-  // Verify expected symbols are present
-  SimpleTestFixture::AssertContainsSymbols(
-      *index, {"test_module", "signal", "byte_t", "init_block"});
-
-  // Verify definition ranges are valid for key symbols
-  for (const auto& [location, info] : all_symbols) {
-    std::string name(info.symbol->name);
-    if (name == "test_module" || name == "signal" || name == "byte_t" ||
-        name == "init_block") {
-      REQUIRE(info.is_definition);
-      REQUIRE(info.definition_range.start().valid());
-    }
-  }
-}
-
-TEST_CASE(
-    "SemanticIndex DefinitionIndex-compatible API basic functionality",
-    "[type_handling]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    module test_module;
-      logic signal;
-      typedef logic [7:0] byte_t;
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-
-  // Test reference storage API
-  const auto& references = index->GetReferences();
-  const auto& all_symbols = index->GetAllSymbols();
-
-  // Basic sanity checks - should have some data
-  REQUIRE(!all_symbols.empty());
-
-  // Verify references are accessible via GetReferences()
-  (void)references;  // May be empty for single-file tests
-
-  // Test that symbols have definition ranges in their SymbolInfo
-  // Verify at least one symbol has valid definition ranges
-  bool found_symbol_with_range = false;
-  for (const auto& [loc, info] : all_symbols) {
-    if (info.is_definition && info.location.valid()) {
-      found_symbol_with_range = true;
-      // Basic check that definition_range is set
-      REQUIRE(info.location.valid());
-      break;
-    }
-  }
-  REQUIRE(found_symbol_with_range);
-}
-
-TEST_CASE(
-    "SemanticIndex collects functions and tasks correctly", "[type_handling]") {
-  SimpleTestFixture fixture;
-  std::string code = R"(
-    module test_module;
-      // Function with explicit return type
-      function automatic logic simple_func();
-        simple_func = 1'b0;
-      endfunction
-
-      // Simple task
-      task automatic simple_task();
-        $display("test");
-      endtask
-    endmodule
-  )";
-
-  auto index = fixture.CompileSource(code);
-  auto symbols = index->GetDocumentSymbols(GetTestUri());
-
-  REQUIRE(!symbols.empty());
-  REQUIRE(symbols[0].children.has_value());
-
-  // Find functions and tasks in module
-  SimpleTestFixture::AssertDocumentSymbolExists(
-      symbols, "simple_func", lsp::SymbolKind::kFunction);
-  SimpleTestFixture::AssertDocumentSymbolExists(
-      symbols, "simple_task", lsp::SymbolKind::kFunction);
-
-  // Verify function is a leaf node (no children shown in document symbols)
-  auto function_symbol = std::find_if(
-      symbols[0].children->begin(), symbols[0].children->end(),
-      [](const auto& s) { return s.name == "simple_func"; });
-  REQUIRE(
-      (!function_symbol->children.has_value() ||
-       function_symbol->children->empty()));
-
-  auto task_symbol = std::find_if(
-      symbols[0].children->begin(), symbols[0].children->end(),
-      [](const auto& s) { return s.name == "simple_task"; });
-  // Tasks should be leaf nodes (no children shown in document symbols)
-  REQUIRE(
-      (!task_symbol->children.has_value() || task_symbol->children->empty()));
-}
-
-}  // namespace slangd::semantic
