+++ test/slangd/services/BUILD.bazel
@@ -2,6 +2,8 @@
 Build file for slangd services tests.
 """
 
+load("@rules_cc//cc:cc_test.bzl", "cc_test")
+
 cc_test(
     name = "overlay_session_test",
     timeout = "short",
